+++ src/client/controllers/BuildPreviewController.ts
@@ -462,6 +462,9 @@ export class BuildPreviewController implements Controller {
       case UnitType.DefensePost:
         rangeRadius = this.game.config().defensePostRange();
         break;
+      case UnitType.Artillery:
+        rangeRadius = this.game.config().artilleryRange();
+        break;
     }
     let radiusTileX = this.game.x(tileRef);
     let radiusTileY = this.game.y(tileRef);