+++ src/client/controllers/SoundEffectController.ts
@@ -224,6 +224,7 @@ export class SoundEffectController implements Controller {
         if (unit.owner() === myPlayer) this.emit("build-port");
         break;
       case UnitType.DefensePost:
+      case UnitType.Artillery:
         if (unit.owner() === myPlayer) this.emit("build-defense-post");
         break;
       case UnitType.SAMLauncher: