+++ src/core/StatsSchemas.ts
@@ -31,6 +31,7 @@ export type BoatUnitType = UnitType.TradeShip | UnitType.TransportShip;
 export const otherUnits = [
   "city",
   "defp",
+  "arty",
   "port",
   "wshp",
   "silo",
@@ -42,6 +43,7 @@ export type OtherUnit = z.infer<typeof OtherUnitSchema>;
 export type OtherUnitType =
   | UnitType.City
   | UnitType.DefensePost
+  | UnitType.Artillery
   | UnitType.MissileSilo
   | UnitType.Port
   | UnitType.SAMLauncher
@@ -51,6 +53,7 @@ export type OtherUnitType =
 export const unitTypeToOtherUnit = {
   [UnitType.City]: "city",
   [UnitType.DefensePost]: "defp",
+  [UnitType.Artillery]: "arty",
   [UnitType.MissileSilo]: "silo",
   [UnitType.Port]: "port",
   [UnitType.SAMLauncher]: "saml",